+++ StripeCore/StripeCore/Source/Localization/String+Localized.swift
@@ -29,5 +29,15 @@ import Foundation
             // THIS STRING SHOULD NOT BE MODIFIED
             STPLocalizedString("We use Stripe to verify your card details. Stripe may use and store your data according its privacy policy. <a href='https://support.stripe.com/questions/stripes-card-image-verification'><u>Learn more</u></a>", "Informational text informing the user that Stripe is used to process data and a link to Stripe's privacy policy")
         }
+
+        public static func scanCardExpectedPrivacyLinkText() -> NSAttributedString? {
+            let stringData = Data(String.Localized.scan_card_privacy_link_text.utf8)
+            let stringOptions: [NSAttributedString.DocumentReadingOptionKey : Any] = [
+                .documentType: NSAttributedString.DocumentType.html,
+                .characterEncoding: String.Encoding.utf8.rawValue
+            ]
+
+            return try? NSAttributedString(data: stringData, options: stringOptions, documentAttributes: nil)
+        }
     }
 }